Add basic implementation of MinimalConcept
This should have been trivial, but I spent more than a day (yesterday
afternoon, this morning,) debugging a typo in a URI that caused a test
to fail.

Apparently, OWL has the C++ behaviour of "if it doesn't exist, create it"
in some cases, so I've written wrappers myself that _do_ throw, like I
was expecting them to when debugging.